[REF] move extract_data (#6351)
