+++ nilearn/_utils/tests/test_common.py
@@ -33,7 +33,7 @@ def test_number_public_functions(matplotlib_pyplot):  # noqa: ARG001
     If this is intentional, then the number should be updated in the test.
     Otherwise it means that the public API of nilearn has changed by mistake.
     """
-    assert len({_[0] for _ in all_functions(return_private=False)}) == 277
+    assert len({_[0] for _ in all_functions(return_private=False)}) == 276
 
 
 def test_number_public_classes(matplotlib_pyplot):  # noqa: ARG001
