+++ nilearn/surface/surface.py
@@ -6,7 +6,6 @@ import pathlib
 import warnings
 from collections.abc import Iterable, Mapping
 from pathlib import Path
-from typing import Any
 
 import numpy as np
 import pandas as pd
@@ -2099,47 +2098,6 @@ def get_data(img, ensure_finite: bool = False) -> np.ndarray:
     if ensure_finite:
         return ensure_finite_data(data)
     return data
-
-
-def extract_data(img, index) -> dict[Any, np.ndarray]:
-    """Extract data of a SurfaceImage a specified indices.
-
-    Parameters
-    ----------
-    img : SurfaceImage object
-
-    index : Any type compatible with numpy array indexing
-        Used for indexing the 2D data array in the 2nd dimension.
-
-    Returns
-    -------
-    a dict where each value contains the data extracted
-    for each part
-    """
-    check_is_of_allowed_type(img, (SurfaceImage,), "img")
-    mesh = img.mesh
-    data = img.data
-    data._check_parts()
-
-    if isinstance(index, np.ndarray):
-        return {hemi: data.parts[hemi][:, index].copy() for hemi in data.parts}
-
-    if isinstance(index, int):
-        last_dim = 1
-    elif isinstance(index, slice):
-        start, stop, step = index.indices(data._n_samples)
-        last_dim = max(0, (stop - start + (step - 1)) // step)
-    elif all(isinstance(x, bool) for x in index):
-        last_dim = sum(index)
-    else:
-        last_dim = len(index)
-
-    return {
-        hemi: data.parts[hemi][:, index]
-        .copy()
-        .reshape(mesh.parts[hemi].n_vertices, last_dim)
-        for hemi in data.parts
-    }
 
 
 def compute_adjacency_matrix(mesh: InMemoryMesh, values="ones", dtype=None):
